[solver] Fix restartGMRes parameter number type

See merge request dumux-repositories/dumux!2631

(cherry picked from commit 9b0db49366fed7183737b59b00f4bcf1e7b10495)

930dd74f [solver] Fix restartGMRes parameter number type

+++ dumux/linear/seqsolverbackend.hh
@@ -1086,7 +1086,7 @@ public:
     {
         BlockDiagILU0Preconditioner<Matrix, Vector, Vector> preconditioner(M);
         Dune::MatrixAdapter<Matrix, Vector, Vector> op(M);
-        static const int restartGMRes = getParamFromGroup<double>(this->paramGroup(), "LinearSolver.GMResRestart");
+        static const int restartGMRes = getParamFromGroup<int>(this->paramGroup(), "LinearSolver.GMResRestart");
         Dune::RestartedGMResSolver<Vector> solver(op, preconditioner, this->residReduction(), restartGMRes,
                                                   this->maxIter(), this->verbosity());
         auto bTmp(b);
